+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.6.0 (unreleased)
 - Add encoding package which contains basic implementations of json and gob encoding
 - Send invoking user's ID when creating a scope
+- Fallback connector fixes, including always return all results as pointers and only writing to fallback when origin succeeds 
 
 ## v2.5.0 (unreleased)